requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
alembic==1.17.2
appnope==0.1.4
asttokens==3.0.1
attrs==25.4.0
audioread==3.1.0
autoflake==2.3.1
autopep8==2.3.2
black==25.12.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorlog==6.10.1
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
datasets==3.6.0
debugpy==1.8.17
decorator==5.2.1
dill==0.3.8
executing==2.2.1
filelock==3.20.0
flake8==7.3.0
fonttools==4.61.0
frozenlist==1.8.0
fsspec==2025.3.0
hf-xet==1.2.0
huggingface-hub==0.36.0
idna==3.11
imbalanced-learn==0.14.1
iniconfig==2.3.0
ipykernel==7.1.0
ipython==9.8.0
ipython-pygments-lexers==1.1.1
ipywidgets==8.1.8
isort==7.0.0
jedi==0.19.2
jinja2==3.1.6
joblib==1.5.2
jupyter-client==8.6.3
jupyter-core==5.9.1
jupyterlab-widgets==3.0.16
kiwisolver==1.4.9
lazy-loader==0.4
librosa==0.11.0
llvmlite==0.46.0
mako==1.3.10
markupsafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
mccabe==0.7.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
multiprocess==0.70.16
mypy-extensions==1.1.0
narwhals==2.13.0
nest-asyncio==1.6.0
networkx==3.6.1
numba==0.63.1
numpy==2.3.5
optuna==4.6.0
packaging==25.0
pandas==2.3.3
parso==0.8.5
pathspec==0.12.1
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.1
plotly==6.5.0
pluggy==1.6.0
pooch==1.8.2
prompt-toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.2
psutil==7.1.3
ptyprocess==0.7.0
pure-eval==0.2.3
pyarrow==22.0.0
pycodestyle==2.14.0
pycparser==2.23
pyflakes==3.4.0
pygments==2.19.2
pyparsing==3.2.5
pytest==9.0.2
python-dateutil==2.9.0.post0
pytokens==0.3.0
pytz==2025.2
pyyaml==6.0.3
pyzmq==27.1.0
regex==2025.11.3
requests==2.32.5
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
setuptools==80.9.0
six==1.17.0
sklearn-compat==0.1.5
soundfile==0.13.1
soxr==1.0.0
sqlalchemy==2.0.45
stack-data==0.6.3
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.9.1
torchaudio==2.9.1
torchcodec==0.9.1
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.3
typing-extensions==4.15.0
tzdata==2025.2
urllib3==2.6.1
wcwidth==0.2.14
widgetsnbextension==4.0.15
xxhash==3.6.0
yarl==1.22.0
